in ScreenWidthContext, make custom hook to return context. Import context into app and wrap app.
